+++ src/react/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useLazyQuery';
 export * from './useMutation';
 export * from './useQuery';
 export * from './useSubscription';
+export * from './useReactiveVar';